requirements.txt
absl-py==2.3.1
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
argparse==1.4.0
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
brax==0.14.0
certifi==2026.1.4
charset-normalizer==3.4.4
chex==0.1.90
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.3
craftax==1.5.0
cycler==0.12.1
decorator==4.4.2
distrax==0.1.5
dm-env==1.6
dm-tree==0.1.8
dotmap==1.3.30
equinox==0.13.4
esquilax==2.1.0
etils==1.13.0
evosax==0.2.0
executing==2.2.1
farama-notifications==0.0.4
filelock==3.20.3
flask==3.1.2
flask-cors==6.0.2
flax==0.10.6
fonttools==4.61.1
fsspec==2026.1.0
gast==0.7.0
gitdb==4.0.12
gitpython==3.1.46
glfw==2.10.0
grpcio==1.78.0
gym==0.26.2
gym-notices==0.1.0
gymnasium==1.2.3
gymnax==0.0.9
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.4.1
humanize==4.15.0
hydra-core==1.3.2
idna==3.11
imageio==2.37.2
imageio-ffmpeg==0.6.0
importlib-resources==6.5.2
ipython==9.9.0
ipython-pygments-lexers==1.1.1
itsdangerous==2.2.0
jax==0.5.3
jax-cuda12-pjrt==0.5.3
jax-cuda12-plugin==0.5.3
jax-tqdm==0.4.0
jax2d==1.0.1
jaxgl==1.0.1
jaxlib==0.5.3
jaxopt==0.8.5
jaxtyping==0.3.7
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.3
jumanji==1.1.1
-e file:///home/eleni/workspace/icml_playground/continual_neuroevolution/source/kinetix
kiwisolver==1.4.9
lxml==6.0.2
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mediapy==1.2.5
ml-collections==1.1.0
ml-dtypes==0.4.1
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.2
mujoco==3.4.1.dev856339887
mujoco-mjx @ git+https://github.com/google-deepmind/mujoco@d86acd7e122bde942e40218e5d837d41cb65b314#subdirectory=mjx
nest-asyncio==1.6.0
networkx==3.6.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cccl-cu12==12.9.27
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
opt-einsum==3.4.0
optax==0.2.6
orbax-checkpoint==0.11.24
packaging==25.0
pandas==3.0.0
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
playground @ git+https://github.com/google-deepmind/mujoco_playground.git@0716a13668f29c06fc9aebfb1cfdfe363e05a315
proglog==0.1.12
prompt-toolkit==3.0.52
protobuf==6.33.4
psutil==7.2.1
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==2.12.5
pydantic-core==2.41.5
pygame==2.6.1
pygame-widgets==1.1.5
pygments==2.19.2
pyopengl==3.1.10
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytinyrenderer==0.0.14
pyyaml==6.0.3
requests==2.32.5
rich==14.2.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
sentry-sdk==2.49.0
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.14.0
tensorboardx==2.6.4
tensorflow-probability==0.25.0
tensorstore==0.1.74
threadpoolctl==3.6.0
toolz==1.1.0
torch==2.9.1
tqdm==4.67.1
trac-optimizer==0.0.2
traitlets==5.14.3
treescope==0.1.10
trimesh==4.11.0
triton==3.5.1
typer==0.22.0
typer-slim==0.22.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
wadler-lindig==0.1.7
wandb==0.24.0
warp-lang==1.11.0
wcwidth==0.2.14
werkzeug==3.1.5
wrapt==2.1.1
zipp==3.23.0
